Trying to come up with a good design.
Trying to use SOLID principles and the PSR's as much as I can. Created an InputValidator that is pretty crude. For better security
it would probably be better if the validation method were improved (maybe overriden in a sub class that extends InputValidator).

Trying to use Factory Pattern and Registry Pattern to decouple as much as possible the creation of specific objects.