Improved the abstraction on Entities.
Entity Settings have 5 properties: "id", "locale", "name", "value" and "type".
To represent settings that have more properties than these 5, there was created
the class EntitySettingWithExtraProperties that extends EntitySettings.

The interface ExtraPropertiesIndicator was created so that the classes that implement it
may indicate whether they have extra properties.

The interface ExtraProperties is used to provide methods for manipulation of the extra properties.