Huge leep forward in the abstraction.
The Entity and EntitySetting objects may now represent any object from the OJS.
There's no need anymore to create classes like Article, ArticleFile, User and so on... as the Entity might represent them.
Nor there is any need to use even the EntitySettingWithExtraProperties as the EntitySetting logic encapsulates it successfully.